check patch before setting home
In this file, I made several changes below:

While setting up the turtles, they are not allowed to overlap, and can not be located on black patches (water).

While homeless turtles are wandering, they are not allowed to stop on black patches (water), they are not allowed to stop on restricted patches (red), an they are not allowed to stop on occupied patches.